Removed unnecessary signals, added slide collision detection code
Some unnecessary signals on the sword attack were removed and the code for the reflection was merged into the main collision code
The player when moving could move into a bullet but it wouldn't count that as being hit, so that has been fixed.
Replaced some consts with vars properly.
Collision layers and masks have changed for projectiles, so they now pass through each other.