Fulled added in the Doodads
There's a bug in Godot that prevents collision boxes from properly
mapping but it can be fixed with polygon colliders. However, that is
more expensive to work with.
As such, if we start getting slowdown from adding doodads, it's for that
reason. Try not to go too ham with them.
Also cleaned up the first two levels to look more presentable as a final
product.